+++ .env.example
@@ -5,5 +5,4 @@ CLIENT_SECRET="123"
 SCOPES="openid myinfo.name myinfo.passport_expiry_date myinfo.nric_number"
 PRIVATE_KEY="-----BEGIN RSA PRIVATE KEY-----woot woot-----END RSA PRIVATE KEY-----"
 OVERRIDE_DEV="http://localhost:3000"
-MOBILE_APP_BUNDLE_ID="exp://192.168.10.37:19000/--"
