+++ .github/workflows/main.yml
@@ -15,7 +15,7 @@ jobs:
     steps:
       - name: Checkout commit
         uses: actions/checkout@v2
-      
+
       - name: Set specified node version
         uses: actions/setup-node@v3
         with:
@@ -23,6 +23,9 @@ jobs:
 
       - name: Install node modules
         run: npm ci
+
+      - name: Build project
+        run: npm run build
 
       # To force hooks to run even during config deployment
       # https://docs.aws.amazon.com/elasticbeanstalk/latest/dg/platforms-linux-extend.html
