+++ .gitignore
@@ -6,3 +6,5 @@ npm-debug.log
 *.pem
 *.der
 test.js
+
+dist
