+++ config/index.ts
@@ -1,0 +1,10 @@
+const BASE_URLS = {
+  prod: 'https://api.id.gov.sg',
+  stag: 'https://api-stg.id.gov.sg',
+  dev: process.env.OVERRIDE_DEV || '',
+}
+const PORT = process.env.PORT || 10000
+const RANDOMNONCE = 'randomnonce'
+const SCOPES = process.env.SCOPES || 'openid'
+
+export { BASE_URLS, PORT, RANDOMNONCE, SCOPES }
