--- lib/config.js
@@ -1,10 +1,0 @@
-const overrideDev = process.env.OVERRIDE_DEV
-const baseUrls = {
-  prod: 'https://api.id.gov.sg',
-  stag: 'https://api-stg.id.gov.sg',
-  dev: overrideDev,
-}
-
-module.exports = {
-  baseUrls,
-}
