--- middlewares/index.js
@@ -1,5 +1,0 @@
-const csp = require('./csp')
-
-module.exports = {
-  csp,
-}
