+++ package.json
@@ -2,26 +2,38 @@
   "name": "sgid-test-app",
   "version": "2.0.0",
   "description": "This is an example node application that implements sgID's OAuth2 API.",
-  "main": "index.js",
+  "main": "dist/index.js",
   "engines": {
     "node": ">=16.16"
   },
   "dependencies": {
     "@opengovsg/sgid-client": "github:opengovsg/sgid-client#test/v2",
+    "@types/uuid": "^9.0.1",
+    "cookie-parser": "^1.4.6",
     "dotenv": "8.6.0",
     "ejs": "3.1.7",
     "express": "4.17.1",
     "helmet": "4.6.0",
     "jsonwebtoken": "8.5.1",
+    "node-cache": "^5.1.2",
     "node-fetch": "2.6.7",
-    "node-jose": "2.1.1"
+    "node-jose": "2.1.1",
+    "ts-node": "^10.9.1",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "nodemon": "^2.0.6"
+    "@types/cookie-parser": "^1.4.3",
+    "@types/express": "^4.17.17",
+    "@types/node": "^20.1.7",
+    "@types/node-cache": "^4.2.5",
+    "nodemon": "^2.0.6",
+    "typescript": "^5.0.4"
   },
   "scripts": {
-    "start": "node server",
-    "dev": "nodemon server"
+    "start": "node -r dotenv/config dist/index",
+    "build": "tsc",
+    "start:build": "npm run build && npm run start",
+    "dev": "nodemon -w . -w .env -r dotenv/config index.ts"
   },
   "repository": {
     "type": "git",
