+++ utils/index.ts
@@ -1,11 +1,8 @@
 /**
  * Formats the data into an array of arrays,
  * specifically for the display on the frontend
- *
- * @param {object} result
- * @returns {array}
  */
-function formatData(result) {
+export const formatData = (result: { [index: string]: string }): string[][] => {
   const formattedResult = []
 
   for (const [key, value] of Object.entries(result)) {
@@ -18,16 +15,9 @@ function formatData(result) {
 /**
  * Converts a key string from dot-delimited into uppercase
  * for frontend display
- *
- * @param {string} key
- * @returns {string}
  */
-function prettifyKey(key) {
+export const prettifyKey = (key: string): string => {
   let prettified = key.split('.')[1]
   prettified = prettified.replace(/_/g, ' ')
   return prettified.toUpperCase()
-}
-
-module.exports = {
-  formatData,
 }
